+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [5.20.0] - 2026-08-04
+
+_Stable release based on [5.20.0-rc.1]._
+
 ## [5.20.0-rc.1] - 2026-08-04
 
 ### Added
@@ -1734,6 +1738,7 @@ _Stable release based on [0.1.0-rc.1]._
 
 ## [0.0.0] - 2024-09-12
 
+[5.20.0]: https://https://github.com/internetguru/laravel-common/compare/v5.19.0...v5.20.0
 [5.20.0-rc.1]: https://github.com/internetguru/laravel-common/releases/tag/v5.19.0
 [5.19.0]: https://https://github.com/internetguru/laravel-common/compare/v5.18.14...v5.19.0
 [5.19.0-rc.1]: https://github.com/internetguru/laravel-common/releases/tag/v5.18.14